Crate re_grpc_client

Communications with an Rerun Data Platform gRPC server.

Modules

address 🔒

message_proxy

Structs

InvalidRedapAddress

The given url is not a valid Rerun storage node URL.

TonicStatusError

Wrapper with a nicer error message

Enums

RedapAddress

Parsed from rerun://addr:port/recording/12345 or rerun://addr:port/catalog

StreamError

Constants

CATALOG_APPLICATION_ID 🔒

CATALOG_BP_STORE_ID 🔒

CATALOG_REC_STORE_ID 🔒

Functions

activate_catalog_blueprint 🔒

spawn_future 🔒

store_info_from_catalog_chunk

stream_catalog_async 🔒

stream_from_redap

Stream an rrd file or metadata catalog over gRPC from a Rerun Data Platform server.

stream_recording_async 🔒